Add CI an release workflows, fix Windows
Turns out the `:raw` encoding layer is not literally raw, but files are
read in on Windows with `\r\n` line endings, not just `\n`. Fixed by
appending the `:crlf` layer to `:raw` to always end up with only `\n`
line endings. This fixes tests that match line endings with `$` in
regular expressions.

Discovered by adding tests for all supported Perls on Linux, macOS, and
Windows, and discovering the test failures on Windows.

Also add a release workflow to release on both CPAN and GitHub, and a
Module::Build subclass to generate the files needed for the GitHub
release.

Update the README with a link to the API docs and using fenced code
blocks.